+++ test/utils.test.js
@@ -903,6 +903,13 @@ describe('utils.printOutdatedSdkWarning()', () => {
         utils.printOutdatedSdkWarning();
     });
 
+    test('should do nothing when APIFY_DISABLE_OUTDATED_WARNING is set', () => {
+        process.env.APIFY_DISABLE_OUTDATED_WARNING = '1';
+        logMock.expects('warning').never();
+        utils.printOutdatedSdkWarning();
+        delete process.env.APIFY_DISABLE_OUTDATED_WARNING;
+    });
+
     test('should correctly work when outdated', () => {
         process.env[ENV_VARS.SDK_LATEST_VERSION] = semver.inc(currentVersion, 'minor');
         console.log(process.env[ENV_VARS.SDK_LATEST_VERSION]);